requirements-v8-optional.txt
# Optional dependencies for v8 diarization (speaker detection)
# Install only if you run cluster_faces_v8.py with --diarize
pyannote.audio
huggingface_hub
librosa
soundfile
omegaconf
